backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
anthropic>=0.28.0
supabase>=2.5.0
python-dotenv>=1.0.0
pydantic-settings>=2.3.0
PyJWT[crypto]>=2.8.0
